+++ app.plugin.js
@@ -2,4 +2,4 @@
  * Expo Config Plugin 진입점
  * 이 파일은 npm 패키지의 루트에 있어야 합니다.
  */
-module.exports = require("./build/plugin.js").default;
+module.exports = require("./build/index.js").default;